JWT + ABMs (CATEGORIAS,CUENTAS, INGRESOS y GASTOS)

+++ src/main/java/com/proyectodesarrollo/gestorfinanzasytareas/services/TareaService.java
@@ -1,9 +1,13 @@
 package com.proyectodesarrollo.gestorfinanzasytareas.services;
 
-import com.proyectodesarrollo.gestorfinanzasytareas.entities.Tarea;
-import com.proyectodesarrollo.gestorfinanzasytareas.repositories.TareaRepository;
+import java.util.List;
+import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import com.proyectodesarrollo.gestorfinanzasytareas.entities.Tarea;
+import com.proyectodesarrollo.gestorfinanzasytareas.repositories.TareaRepository;
 
 @Service
 public class TareaService {
@@ -11,15 +15,34 @@ public class TareaService {
     @Autowired
     private TareaRepository tareaRepository;
 
+    public List<Tarea> getAllTareas() {
+        return tareaRepository.findAll();
+    }
+
+    public Optional<Tarea> getTareaById(Long id) {
+        return tareaRepository.findById(id);
+    }
+
     public Tarea createTarea(Tarea tarea) {
         return tareaRepository.save(tarea);
     }
 
-    public Tarea getTareaById(Long id) {
-        return tareaRepository.findById(id).orElse(null);
+    // Actualizar una tarea existente
+    public Tarea updateTarea(Long id, Tarea tareaActualizada) {
+        return tareaRepository.findById(id)
+                .map(tarea -> {
+                    tarea.setTitulo(tareaActualizada.getTitulo());
+                    tarea.setDescripcion(tareaActualizada.getDescripcion());
+                    tarea.setCompletada(tareaActualizada.isCompletada());
+                    tarea.setCuenta(tareaActualizada.getCuenta());
+                    return tareaRepository.save(tarea);
+                }).orElseThrow(() -> new RuntimeException("Tarea no encontrada con ID: " + id));
     }
 
     public void deleteTarea(Long id) {
+        if (!tareaRepository.existsById(id)) {
+            throw new RuntimeException("Tarea no encontrada con id: " + id);
+        }
         tareaRepository.deleteById(id);
     }
 }
